+++ HISTORY.rst
@@ -12,6 +12,9 @@ Unreleased Changes
   using custom 2D numpy arrays, such as those commonly used in image processing
   (``pygeoprocessing.kernel_from_numpy_array``).
   https://github.com/natcap/pygeoprocessing/issues/268
+* Logging across functions in ``pygeoprocessing.geoprocessing`` now correctly
+  reports the function that it's being called from rather than ``<lambda>``.
+  https://github.com/natcap/pygeoprocessing/issues/300
 * The function ``pygeoprocessing.reproject_vector`` now accepts an optional
   parameter ``layer_name`` to allow the target vector layer name to be defined
   by the user.  If the user does not provide a ``layer_name``, the layer name